🚑  Packing

+++ contracts/contracts/lab/sales/public/Data.sol
@@ -1,8 +1,6 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.0;
 
-import {Packer} from "../../../lib/Packer.sol";
-
 import {Clone} from "clones-with-immutable-args/Clone.sol";
 
 import {ERC721} from "@rari-capital/solmate/src/tokens/ERC721.sol";
@@ -11,12 +9,18 @@ import {ERC20} from "@rari-capital/solmate/src/tokens/ERC20.sol";
 import {GBCLab as Lab} from "../../Lab.sol";
 
 contract PublicData is Clone {
+    /// @notice Chunk1
+    /// @return lab             : address
+    /// @return wallet          : uint96
     function _chunk1() private pure returns (Lab lab, uint96 wallet) {
         uint256 i = _getArgUint256(0);
         lab = Lab(address(uint160(i)));
         wallet = uint96(i >> 160);
     }
 
+    /// @notice Chunk2
+    /// @return receiver        : address
+    /// @return transaction     : uint96
     function _chunk2()
         private
         pure
@@ -27,12 +31,19 @@ contract PublicData is Clone {
         transaction = uint96(i >> 160);
     }
 
+    /// @notice Chunk3
+    /// @return token           : address
+    /// @return finish          : uint96
     function _chunk3() private pure returns (ERC20 token, uint96 finish) {
         uint256 i = _getArgUint256(64);
         token = ERC20(address(uint160(i)));
         finish = uint96(i >> 160);
     }
 
+    /// @notice Chunk4
+    /// @return start           : uint64
+    /// @return supply          : uint96
+    /// @return cost            : uint96
     function _chunk4()
         private
         pure
@@ -48,6 +59,8 @@ contract PublicData is Clone {
         cost = uint96(i >> 160);
     }
 
+    /// @notice Chunk5
+    /// @return item         : uint256
     function _chunk5() private pure returns (uint256) {
         return _getArgUint256(160);
     }
@@ -94,9 +107,9 @@ contract PublicData is Clone {
             uint96 transaction,
             ERC20 token,
             uint96 finish,
-            uint96 start,
-            uint128 supply,
-            uint128 cost,
+            uint64 start,
+            uint96 supply,
+            uint96 cost,
             uint256 item
         )
     {
